requirements/requirements-base.txt
Django==1.7.4
django-bootstrap3==5.1.1
PyYAML==3.11
django-localflavor==1.1